docs: remove net._http_collect_response usage
The _http_collect_response was never meant to be used by end users. It
only was meant as an internal function for testing.

Why? Because it uses an endless loop (net._await_response) which is not
safe when running inside a transaction.

Previously it was removed from

https://supabase.com/docs/guides/database/extensions/pg_net

For the same reason.

If there's a need for sync requests, the current recommendation
is to use pgsql-http.