refactor: remove strict and parallel safe attrs
Closes https://github.com/supabase/pg_net/issues/172.

Remove strict as explained on https://github.com/supabase/pg_net/pull/172.
A test is added on a previous commit to ensure the behavior is
maintained.

Also, remove PARALLEL SAFE since these functions are really PARALLEL UNSAFE,
as mentioned on the docs:

"Functions and aggregates must be marked PARALLEL UNSAFE if they write to the database,"
https://www.postgresql.org/docs/current/parallel-safety.html

Also remove VOLATILE since it's implicit.