feat: process the queue on demand
Avoid polling the queue table and instead wait for a `wake` signal that
is sent whenever the `net.http_get` and friends are called.
The worker will be "awake" as long as the request queue has rows in it or as long there are expired responses.

If the worker is restarted while it is "awake", it will keep
processing the queue and expired responses once it starts again.

This causes some behaviors tochange:

- `pg_net.ttl`: Previously the TTL
was fulfilled a few seconds away but now it will only happen while the
worker is awake, it's not ensured to happen soon after the TTL expires.
The `ttl` is still honored though and its raison d'être is still
maintained, it will prevent the response table from growing too much.

- `net.http_request_queue`: direct insertions will no longer cause new
  requests to fire. This should be fine since this table was never meant
  to be used this way, users should only rely on the `net` schema
  functions to queue new requests.

Closes https://github.com/supabase/pg_net/issues/164, no
unnecessary queries are now done and we don't pollute
`pg_stat_statements`.

Closes https://github.com/supabase/pg_net/issues/175, requests will
no longer wait one second all the time. The queue will be processed as
soon as new requests arrive.