fix: locking the tables for too much time
If the worker process the queue for a long time, it will
preventing operations like TRUNCATE or VACUUM FULL from completing.

This is because it acquires an AccessShareLock on the extension tables
and only releases it once it goes to sleep.

Now we use `ConditionalLockRelationOid` and unlock the tables
at the end of each transaction, allowing VACUUM FULL or TRUNCATE to
finish fast.