refactor: remove unnecessary memory context switches
Currently memory contexts are used to work around the short memory
context of SPI blocks (SPI_connect and SPI_finish), which are used
inside some functions. For example:

https://github.com/supabase/pg_net/blob/314d00e34e22378ca2b733838f96b4c179d727a9/src/core.c#L139-L170

https://github.com/supabase/pg_net/blob/314d00e34e22378ca2b733838f96b4c179d727a9/src/core.c#L172-L227

But it's unnecessary to do that, instead we can use a single SPI block
in the worker loop and share the same memory context for the curl
operations.

This makes the code simpler and makes memory allocations more visible.